+++ src/ApiSettings.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Rhubarb\Scaffolds\TokenBasedRestApi;
+
+use Rhubarb\Crown\Settings;
+
+class ApiSettings extends Settings
+{
+    public $tokenExpiration = '+1 day';
+    public $extendTokenExpirationOnUse = false;
+}