+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+### 1.1.1
+* Added:    ApiSettings class
+* Changed:  Token expiration time is pulled from settings. Old constant deprecated
+* Added:    Support for extending token expiration on use
+
 ### 1.1.0
 
 * Added:	Changelog